Added /register and /login
Added register and login POST endpoints. After registration/login, user is provided with access and refresh tokens. Additional work (various endpoints etc.) has to be done on the token part (refreshing, invalidating etc.). PROBLEM: /login endpoint receives plain-text password. If it receives hashed password, there is no way to check whether the password is correct. If we check the password on the client (Android app) side, there is no way to know when to send the tokens. Closes #1, closes #2, work on #3.