Updated JWT (in)validation
Switched from previous commit's problematic hashing approach to a simpler one. Last logout timestamp is saved to the DB each time the user logs out. Token is identified as valid only if his issue timestamp is older than both user's creation timestamp and user's last logout timestamp. Potential problem: unauthorized modification of last logout timestamp. (maybe use encryption with a secret key?) Access JWT needs to be added to each info request to this API that is user-specific from now on (so, first check whether the access token is valid (if not, send a feedback saying that it needs to be refreshed via /accessToken) and if it is valid, send the requested info as a response). Closes #3. Closes #4.